+++ components/Settings/AutomaticNextLeg.vue
@@ -65,20 +65,11 @@
 <script setup lang="ts">
 import AppToggle from "../AppToggle.vue";
 import AppInput from "../AppInput.vue";
-import { AutodartsToolsConfig, type IConfig, updateConfigIfChanged } from "@/utils/storage";
+
+import { AutodartsToolsConfig, type IConfig, defaultConfig } from "@/utils/storage";
 
 const emit = defineEmits([ "toggle", "settingChange" ]);
 const config = ref<IConfig>();
-
-onMounted(async () => {
-  config.value = await AutodartsToolsConfig.getValue();
-});
-
-watch(config, async () => {
-  const currentConfig = await AutodartsToolsConfig.getValue();
-  await nextTick();
-  await updateConfigIfChanged(currentConfig, config.value, "automaticNextLeg");
-}, { deep: true });
 
 async function toggleFeature() {
   if (!config.value) return;
@@ -93,4 +84,17 @@ async function toggleFeature() {
     emit("toggle", "automatic-next-leg");
   }
 }
+
+onMounted(async () => {
+  config.value = await AutodartsToolsConfig.getValue();
+  if (!config.value?.automaticNextLeg) config.value.automaticNextLeg = defaultConfig.automaticNextLeg;
+});
+
+watch(config, async (_, oldValue) => {
+  if (!oldValue) return;
+
+  await AutodartsToolsConfig.setValue(toRaw(config.value!));
+  emit("settingChange");
+  console.log("Automatic Next Leg setting changed");
+}, { deep: true });
 </script>